Moving get_date function to anhalyze module.
